+++ src/main/java/carracing/domain/car/Car.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.domain.car;
 
 public class Car {
