+++ src/main/java/carracing/domain/race/move/MoveStrategy.java
@@ -1,4 +1,4 @@
-package carracing.move_strategy;
+package carracing.domain.race.move;
 
 public interface MoveStrategy {
     int getMovePoint();
