+++ src/test/java/carracing/CarTest.java
@@ -1,5 +1,6 @@
 package carracing;
 
+import carracing.domain.car.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
